split db and index
