+++ db/index.js
@@ -1,0 +1,7 @@
+const { Pool } = require('pg');
+
+const pool = new Pool();
+
+module.exports = {
+    query: (text, params) => pool.query(text, params)
+};
